rft(*): reorganize code with proper file name
