+++ benchmark/client/main.go
@@ -31,6 +31,7 @@ import (
 
 import (
 	"github.com/dubbogo/gost/sync"
+
 	"github.com/montanaflynn/stats"
 )
